[2287] Amended user jounery
